+++ source/TrickHLA/utils/SleepTimeout.cpp
@@ -30,7 +30,6 @@ NASA, Johnson Space Center\n
 
 // Trick includes.
 #include "trick/clock_proto.h"
-#include "trick/exec_proto.h"
 
 // TrickHLA includes.
 #include "TrickHLA/utils/SleepTimeout.hh"
@@ -106,11 +105,8 @@ int SleepTimeout::sleep() const
 // Current time as an integer in microseconds.
 int64_t SleepTimeout::time()
 {
-   // Make sure the time is always an integer in microseconds.
-   int const time_tic_value = exec_get_time_tic_value();
-   return ( time_tic_value == 1000000 )
-             ? clock_wall_time()
-             : ( ( clock_wall_time() * 1000000 ) / time_tic_value );
+   // This Trick clock API is always an integer in microseconds.
+   return clock_wall_time();
 }
 
 bool SleepTimeout::timeout() const